% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AyaJuliapackage_final.R
\name{plot_histogram_residuals}
\alias{plot_histogram_residuals}
\title{Plot Histogram of Residuals}
\usage{
plot_histogram_residuals(residuals)
}
\arguments{
\item{residuals}{A \code{vector} representing the residuals.}
}
\value{
A \code{graph} with the residuals.
}
\description{
This function will plot the histogram for the residuals for the my_lm function.
}
\examples{
variables <- matrix(rnorm(400), ncol = 4)
response <- rnorm(100)
alpha <- 0.05

result <- my_lm(response, variables, alpha)
plot_histogram_residuals(result$residuals)
}
\author{
AyaJulia
}
