% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AyaJuliapackage_final.R
\name{plot_residual}
\alias{plot_residual}
\title{Plot Residuals}
\usage{
plot_residual(y, residuals)
}
\arguments{
\item{y}{A \code{vector} representing the response.}

\item{residuals}{A \code{vector} representing the residuals.}
}
\value{
A \code{graph} with the residuals.
}
\description{
This function will plot the residuals for the my_lm function.
}
\examples{
variables <- matrix(rnorm(400), ncol = 4)
response <- rnorm(100)
alpha <- 0.05

result <- my_lm(response, variables, alpha)
plot_residuals(y, result$residuals)
}
\author{
AyaJulia
}
