% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AyaJuliapackage_final.R
\name{plot_residuals_fitted}
\alias{plot_residuals_fitted}
\title{Plot Residuals versus fitted values}
\usage{
plot_residuals_fitted(predictor, residuals)
}
\arguments{
\item{predictor}{A \code{vector} representing the predictor.}

\item{residuals}{A \code{vector} representing the residuals.}
}
\value{
A \code{graph} with the residuals.
}
\description{
This function will plot the residuals x the fitted values for the residuals for the my_lm function.
}
\examples{
variables <- matrix(rnorm(400), ncol = 4)
response <- rnorm(100)
alpha <- 0.05

result <- my_lm(response, variables, alpha)
plot_residuals_fitted(result$predictor,result$residuals)
}
\author{
AyaJulia
}
